removed lava flow statmanager removed std cout removed lonely growing array include removed binary_tree (????) include from text fixed some signed/unsigned mismatches

+++ Source/GUI/MoneyWidget.cpp
@@ -37,7 +37,7 @@ namespace GUI
 	{
 		SUPRESS_UNUSED_WARNING(aDeltaTime);
 
-		short cheapestHatPrice = GetCheapestHatPrice();
+		unsigned int cheapestHatPrice = GetCheapestHatPrice();
 		if (cheapestHatPrice == 0)
 		{
 			cheapestHatPrice = PollingStation::playerData->GetGold();
